Eliminating some advanced conversion and shadow declaration compiler warnings.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@1843430 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/http2/h2_proxy_session.c
@@ -1423,7 +1423,7 @@ run_loop:
                 ap_log_cerror(APLOG_MARK, APLOG_TRACE3, status, session->c, 
                               APLOGNO(03365)
                               "h2_proxy_session(%s): WAIT read, timeout=%fms", 
-                              session->id, (float)session->wait_timeout/1000.0);
+                              session->id, session->wait_timeout/1000.0);
                 if (status == APR_SUCCESS) {
                     have_read = 1;
                     dispatch_event(session, H2_PROXYS_EV_DATA_READ, 0, NULL);
